+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - Bugfix: Fixed suspicious user treatment update messages not being searchable. (#5865)
 - Dev: Highlight checks now use non-capturing groups for the boundaries. (#5784)
 - Dev: Updated Conan dependencies. (#5776)
+- Dev: Disable QT keywords (i.e. `emit`, `slots`, and `signals`). (#5882)
 - Dev: Replaced usage of `parseTime` with `serverReceivedTime` for clearchat messages. (#5824, #5855)
 - Dev: Support Boost 1.87. (#5832)
 - Dev: Words from `TextElement`s are now combined where possible. (#5847)